[FLINK-10309][rest] Before shutting down cluster, wait for asynchronous operations.
Wait for the result of asynchronous operations to be served before shutting down
the cluster.  This is necessary for the "cancel with savepoint" operation. If we
do not wait for the result to be accessed by the client, we may shutdown the
cluster, and the client gets a ConnectionException.

Extract CompletedOperationCache from AbstractAsynchronousOperationHandlers to
ease unit testing.

This closes #6785.