[hotifx][runtime] Remove NonRecoverableError annotation from NoResourceAvailableException
With new generation scheduler, NonRecoverableError will suppress restarts and
cause jobs to fail. Thus NoResourceAvailableException with NonRecoverableError
annotation will lead jobs to be FAILED, which did not happen in the past. This
should not happen because most issues that causes NoResourceAvailableException
can be recovered via retries/restarts. To suppress restarts by purpose, one
should pass a SuppressRestartsException as the param when creating a
NoResourceAvailableException.